refactor: 공연 업데이트 로직 수정
